put the modal call from page into the card component
